Update serverVersion validation in OpRejectedSchema to enforce integer values
- Modified the serverVersion field in OpRejectedSchema to require integer values, enhancing data integrity and validation for rejected operations.

+++ packages/validation/src/index.ts
@@ -107,7 +107,7 @@ export const OpRejectedSchema = z.object({
   kind: z.literal("op_rejected"),
   op: z.literal(OPS_NAMES.UPDATE),
   reason: z.literal("stale_version"),
-  serverVersion: z.number().min(0),
+  serverVersion: z.number().int().min(0),
   shape: ShapeSchema,
 });
